Beginning new plugin for CVSAnaly
GlueTheos, using the "revision approach"

First version, it just retrieves every pair file-revision, calculates SHA1 and Nilsimsa, and stores all in the database (along with the file itself, compressed with gzip). With the file in the database, the CVS server is no longer necessary in order to obtain the files to be measured.