Fixed graphs plugin, that did not work with the change of the "inAttic" field by the field "state".
Anyway, I have removed also the field "state" by a more clear field called "removed".

The "inAttic" field provided wrong information. Every time a file was found in the attic of the repository, all the revisions (even those where the file was only modified) were marked as "inAttic". Therefore, the file appeared in the whole history as it would be in the attic.

The field "state" was simply the content of the "state" field of the log ("Exp" if the revision is modification, "dead" if it is a deletion).

I have replaced that field by a field called "removed", which is of bool type, and it is True if the file was deleted in that (and only in that) revision. All the previous revisions of a file remain as "False" in the removed field.

I have reviewed all the plugins to make sure that all of them use this new field to locate deleted files and revisions. I have changed the names of some variables and methods accordingly.

I have tested the graphs plugin with the CVS of NetBSD, and seems to work.

Please remember that all the previous information that used the "inAttic" field was wrong, because a lot of revisions that were not in the attic were marked as being in the attic.