Add ICursor class to split up queries
The result of a query is always stored in memory, which means that
queries that return a large amount of data might fill up the memory.
ICursor is a class that uses a real cursor emulating the execute +
fetchmany API, but splitting up the query by using limit, so that the
query is executed with different limit values everytime a fetchmany is
required.