Refactored the code that get the id of a tag
Fixed the cache support for tags.  It was not storing any value
into the cache, hence there was always asking for values to the
database.

In order to improve the encapsulation, the method __get_tag was
added, which is in charge of getting the proper either from the
database or cache, and update the cache.