Refactored the code that get the id of a person
In order to improve the encapsulation, the method __ensure_person
was moved inside of __get_person (a new method added), which is
in charge of getting the proper person id either from the database
or cache, and update the cache.

As a side effect, the code of DBContentHandler::commit() gets a
bit smaller.