[git] Fix error parsing logs that do not include info about branches
CVSAnalY does not parse logs in which its top commit do not include
information about the branches where it belongs to. That is a typical
problem when parsing log files or local repositories.

This patch fixes this issue adding top and following commits,
thouse that do not provide info about their branches, to "(no-branch)"
branch.